old-tests/submission/Microsoft/attribute/anchor_href.htm

<!DOCTYPE html>
<head>
<title>HTMLAnchorElement href Attribute Test</title>
    <link rel="help" href="http://www.w3.org/TR/html5/the-a-element.html#dom-a-href" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body onload="runTests()">

<script>
// The anchor tags used for the test are at the bottom of the page
// The table below contains the tests
// For each row in the test, test_href is called, which maps to the tag, description and value

setup({explicit_done: true});

function test_href(testnumber)
{

    description = document.getElementById('test_' + testnumber + '_description').innerHTML;
    test(function () {
	    var oAnchor     = document.getElementById('anchor_' + testnumber);
	    var value       = document.getElementById('test_' + testnumber + '_value').innerHTML;
	    // null needs to be special cased, same with an empty string
	    if( value == 'null')
		{
		    value = null;
		}
	    else if( value == 'empty string')
		{
		    value = '';
		}

	    assert_equals(oAnchor.getAttribute('href'), value);

	    oAnchor.setAttribute('href', '');
	    assert_equals(oAnchor.getAttribute('href'), '');

	    // remove the attribute for null
	    if (value == null)
		{
		    oAnchor.removeAttribute('href');
		}
	    else
		{
		    oAnchor.setAttribute('href', value);
		}
	    assert_equals(oAnchor.getAttribute('href'), value);
	}, description);
}

function runTests()
{
    var iTests = document.getElementById('testtable').rows.length;
    
    for (var iTest=1;iTest<iTests;iTest++)
	{
	    test_href(iTest);
	}
    done();
}
</script>

<table id='testtable' border='1'>
<tbody>
<tr>
<td>Test Description</td><td>Test Value</td>
</tr>
<tr>
 <td id='test_1_description'>Test get/set on a relative href attribute</td>
 <td id='test_1_value'>anchor.htm</td>
</tr>
<tr>
 <td id='test_2_description'>Test get/set on a absolute href attribute</td>
 <td id='test_2_value'>http://somesite.com/anchor.htm</td>
</tr>
<tr>
 <td id='test_3_description'>Test href attribute is an empty string</td>
 <td id='test_3_value'>empty string</td>
</tr>
<tr>
 <td id='test_4_description'>Test for non-existent href attribute</td>
 <td id='test_4_value'>null</td>
</tr>
<tr>
 <td id='test_5_description'>Test that a href attribute contains search string</td>
 <td id='test_5_value'>http://www.anothersite.com/path/page.htm?parameter=this%20is%a%20parameter</td>
</tr>
<tr>
 <td id='test_6_description'>Test that a href attribute contains the username/password and the search string</td>
 <td id='test_6_value'>http://username:password@www.anothersite.com/path/page.htm?parameter=this%20is%a%20parameter</td>
</tr>
<tr>
 <td id='test_7_description'>Test that a href attribute has not been lowercased</td>
 <td id='test_7_value'>http://WWW.SomeSite.com/INDEX.html</td>
</tr>
</tbody>
</table>

<div id="log"></div>

<a id='anchor_1' href='anchor.htm'>Simple Anchor Relative Path</a><br>
<a id='anchor_2' href='http://somesite.com/anchor.htm'>Simple Anchor Non-Relative Path</a><br>
<a id='anchor_3' href=''>Empty String Anchor</a><br>
<a id='anchor_4'>Anchor with no attributes</a><br>
<a id='anchor_5' href='http://www.anothersite.com/path/page.htm?parameter=this%20is%a%20parameter'>href with a search string that is urlencoded</a><br>
<a id='anchor_6' href='http://username:password@www.anothersite.com/path/page.htm?parameter=this%20is%a%20parameter'>href with a username, password and search string that is urlencoded</a><br>
<a id='anchor_7' href='http://WWW.SomeSite.com/INDEX.html'>Simple Anchor with some uppercase letters</a><br>

</body>
</html>
